core: Move DataAccessTokenManager to cellbase-core #TASK-3842
